artwork follow unfollow feature

+++ components/reusable/AboutCurator.jsx
@@ -1,11 +1,15 @@
 import React, { useEffect, useState } from "react";
 import Link from "next/link";
 import Image from "next/image";
-import { getArtistRecognitions } from "lib/backend";
+import { getArtistRecognitions, unfollowArtist, isFollowed, postFollower, getUserMe } from "lib/backend";
+import Loader from "../svg/Loader";
 
 const AboutCurator = ({ owner }) => {
 
   const [recognitions, setRecognitions] = useState([]);
+  const [following, setFollowing] = useState(false)
+  const [loading, setLoading] = useState(false)
+  const [authUser, setAuthUser] = useState({})
 
   // Owner name to string
   let displayName;
@@ -19,6 +23,24 @@ const AboutCurator = ({ owner }) => {
     displayName = owner.email;
   }
 
+  const handleFollowUnfollow = (isFollow) => {
+    setLoading(true)
+    const token = localStorage.getItem('token')
+    if(isFollow) {
+      //unfollow code
+      unfollowArtist(token, { user_id: owner.id })
+        .then((res) => setFollowing(false))
+        .catch(err => console.log(err))
+        .finally(() => setLoading(false))
+    } else {
+      // follow code
+      postFollower(token, {user_id: owner.id})
+        .then((res) => setFollowing(true))
+        .catch(err => console.log(err))
+        .finally(() => setLoading(false))
+    }
+  }
+
   const getRecognitions = async(artist_id)  => {
     const data = await getArtistRecognitions(artist_id);
     setRecognitions(data);
@@ -26,6 +48,21 @@ const AboutCurator = ({ owner }) => {
   useEffect(() => {
     if(owner.id){
       getRecognitions(owner.id)
+      const token = localStorage.getItem('token')
+      setLoading(true)
+      isFollowed(token, owner.id)
+        .then((res) => setFollowing(res.data))
+        .catch((err) => {
+          console.log(err)
+          setFollowing(false)
+        })
+        .finally(() => setLoading(false))
+
+      
+      getUserMe(token)
+        .then(res => setAuthUser(res))
+        .catch(err => console.log(err))
+        
     }
   },[owner]);
 
@@ -48,7 +85,7 @@ const AboutCurator = ({ owner }) => {
         <div className="w-[55%] pt-[100px] md:pt-0 pr-[15px] pb-10 ml-auto md:ml-0 md:absolute bottom-10 right-10  h-fit">
           <div className="flex items-end gap-2 flex-nowrap">
             <p className="b3">By</p>
-            <p className="truncate l2">{displayName}</p>
+            <Link href={`/people/${owner.id}`}><p className="truncate l2">{displayName}</p></Link>
           </div>
 
           <h4 className="mt-5 mb-10 b2 md:h2">{owner.description}</h4>
@@ -65,11 +102,22 @@ const AboutCurator = ({ owner }) => {
               {owner?.walletAddress?.slice(0,4).toLowerCase()}...{owner?.walletAddress?.slice(-4).toLowerCase()}
             </p>
           </div>
-          <Link href={`/people/${owner.id}`}>
-            <button className="btn btn-secondary hover:bg-[#292928] btn-full border-unveilWhite">
-              Follow artist
-            </button>
-          </Link>
+          { 
+            owner.id !== authUser.id && (
+              <button 
+                className="btn btn-secondary hover:bg-[#292928] btn-full border-unveilWhite"
+                onClick={() => handleFollowUnfollow(following)}
+              >
+                {
+                  loading ? (
+                    <div className="h-[25px] animate-spin justify-center flex items-center">
+                      <Loader color="#F7F4ED" />
+                    </div>
+                  ) : following ? `Following` : `Follow artist`
+                }
+              </button>
+            )
+          }
         </div>
       </div>
     </section>
